Fix AES-GCM dissection
This was producing "None", "None" for the authentication.  I first
changed that to "NULL", "" to be actual valid choices.

However, Wireshark would still fail to properly dissect the AES-GCM
encrypted packets, as it could not find the "Next header".  This was
because it did not know the size of the ICV at the end of the packet.
Wireshark (at least the version I'm using) seems to not know how to
verify AES-GCM authentication.  But we can tell it the size of the ICV
and it will then be able to find the Next header field and properly
dissect the packet inside the ESP.

The only one I've actually tested is the 128 bit scenario.